# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfirmInput": {
        "properties": {
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "model_overrides": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Overrides"
          },
          "push_to_github": {
            "default": false,
            "title": "Push To Github",
            "type": "boolean"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "ConfirmInput",
        "type": "object"
      },
      "DagInput": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "model_overrides": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Overrides"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "DagInput",
        "type": "object"
      },
...
      "ExecRetryInput": {
        "description": "Body for POST /exec/retry — fix-list #14.\n\n`job_id` kept as str (not UUID) to preserve current 400 'Invalid job_id\nformat' error on malformed UUIDs. Pydantic UUID type would return 422\ninstead → would break any client parsing the 400 shape.",
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "node_key": {
            "title": "Node Key",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "node_key"
        ],
        "title": "ExecRetryInput",
        "type": "object"
      },
      "ExecuteNextInput": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "model_overrides": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Overrides"
          },
          "skip_optimize": {
            "default": false,
            "title": "Skip Optimize",
            "type": "boolean"
          },
          "skip_verify": {
            "default": false,
            "title": "Skip Verify",
            "type": "boolean"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "ExecuteNextInput",
        "type": "object"
      },
      "ExecutionResult": {
        "properties": {
          "awaiting_approval": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awaiting Approval"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "model_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Used"
          },
          "node_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Key"
          },
          "output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "prompt_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Used"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "tool": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool"
          },
          "verification_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Reason"
          },
          "verified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified"
          }
        },
        "required": [
          "status"
        ],
        "title": "ExecutionResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IdeaInput": {
        "properties": {
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "idea": {
            "title": "Idea",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "model_overrides": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Overrides"
          }
        },
        "required": [
          "idea"
        ],
        "title": "IdeaInput",
        "type": "object"
      },
...
      "PromptOptimizeInput": {
        "properties": {
          "model_optimizer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Optimizer"
          },
          "model_overrides": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Overrides"
          },
          "model_verifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Verifier"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "skip_verify": {
            "default": false,
            "title": "Skip Verify",
            "type": "boolean"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "PromptOptimizeInput",
        "type": "object"
      },
      "PromptOptimizeResult": {
        "properties": {
          "clarity_score": {
            "title": "Clarity Score",
            "type": "number"
          },
          "intent_preserved": {
            "title": "Intent Preserved",
            "type": "boolean"
          },
          "issues_found": {
            "items": {
              "type": "string"
            },
            "title": "Issues Found",
            "type": "array"
          },
          "issues_resolved": {
            "items": {
              "type": "string"
            },
            "title": "Issues Resolved",
            "type": "array"
          },
          "model_used": {
            "title": "Model Used",
            "type": "string"
          },
          "optimized_prompt": {
            "title": "Optimized Prompt",
            "type": "string"
          },
          "original_prompt": {
            "title": "Original Prompt",
            "type": "string"
          },
          "pre_cleaned": {
            "title": "Pre Cleaned",
            "type": "string"
          },
          "token_count_after": {
            "title": "Token Count After",
            "type": "integer"
          },
          "token_count_before": {
            "title": "Token Count Before",
            "type": "integer"
          },
          "token_reduction_pct": {
            "title": "Token Reduction Pct",
            "type": "number"
          },
          "verifier_used": {
            "title": "Verifier Used",
            "type": "string"
          }
        },
        "required": [
          "original_prompt",
          "optimized_prompt",
          "pre_cleaned",
          "token_count_before",
          "token_count_after",
          "token_reduction_pct",
          "clarity_score",
          "intent_preserved",
          "issues_found",
          "issues_resolved",
          "model_used",
          "verifier_used"
        ],
        "title": "PromptOptimizeResult",
        "type": "object"
      },
...
      "SkipNodeInput": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "node_key": {
            "title": "Node Key",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "node_key"
        ],
        "title": "SkipNodeInput",
        "type": "object"
      },
...
      "StatusResponse": {
        "properties": {
          "recent_jobs": {
            "items": {
              "$ref": "#/components/schemas/app__routers__status__JobSummary"
            },
            "title": "Recent Jobs",
            "type": "array"
          },
          "status_counts": {
            "$ref": "#/components/schemas/StatusCounts"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "total_jobs": {
            "title": "Total Jobs",
            "type": "integer"
          }
        },
        "required": [
          "status_counts",
          "total_jobs",
          "recent_jobs",
          "timestamp"
        ],
        "title": "StatusResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Self-hosted RAG-powered workflow orchestrator",
    "title": "Scaffold Engine",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/dag": {
      "post": {
        "description": "Step 11: Generate DAG from refined idea brief.",
        "operationId": "generate_dag_endpoint_dag_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DagInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Generate Dag Endpoint"
      }
    },
    "/dag/{job_id}": {
      "get": {
        "description": "Step 18: Retrieve DAG nodes + job status for a job.",
        "operationId": "get_dag_dag__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Dag"
      }
    },
...
    "/exec/retry": {
      "post": {
        "description": "Reset a failed node to pending for retry.",
        "operationId": "exec_retry_exec_retry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecRetryInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Exec Retry"
      }
    },
    "/exec/status/{job_id}": {
      "get": {
        "description": "Get execution state for a job.",
        "operationId": "exec_status_exec_status__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Exec Status"
      }
    },
    "/execute": {
      "post": {
        "description": "Step 15: Execute the next pending DAG node for a job.\n\nNo DB dependency: execute_next_node manages its own short-lived sessions.",
        "operationId": "execute_next_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteNextInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Execute Next",
        "tags": [
          "Step 15"
        ]
      }
    },
    "/execute/all": {
      "post": {
        "description": "Execute all DAG nodes in sequence, streaming SSE events.\nAuto-generates DAG if none exists.  Failed nodes are skipped;\ndownstream nodes blocked by failures are reported at the end.",
        "operationId": "execute_all_endpoint_execute_all_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteNextInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Execute All Endpoint",
        "tags": [
          "Step 15"
        ]
      }
    },
...
    "/ideas": {
      "post": {
        "description": "Step 10: Submit new idea → trigger refinement.",
        "operationId": "submit_idea_ideas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdeaInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Submit Idea"
      }
    },
    "/ideate": {
      "post": {
        "description": "Phase 1: Analyze idea, assess feasibility, halt for confirmation.",
        "operationId": "ideate_endpoint_ideate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdeaInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Ideate Endpoint"
      }
    },
    "/ideate/confirm": {
      "post": {
        "description": "Phase 2: User confirms -> research -> ingest -> compile -> present workflow.",
        "operationId": "ideate_confirm_endpoint_ideate_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Ideate Confirm Endpoint"
      }
    },
...
    "/optimize": {
      "post": {
        "description": "Step 14: Optimize a prompt — strip filler, reduce tokens, verify intent, score clarity.",
        "operationId": "optimize_endpoint_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptOptimizeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptOptimizeResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Optimize Endpoint",
        "tags": [
          "Step 14"
        ]
      }
    },
...
    "/skip": {
      "post": {
        "description": "Step 15: Skip a specific DAG node.",
        "operationId": "skip_node_endpoint_skip_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkipNodeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Skip Node Endpoint",
        "tags": [
          "Step 15"
        ]
      }
    },
...
    "/status": {
      "get": {
        "description": "Return job status counts and recent jobs.",
        "operationId": "get_status_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "refining",
                    "awaiting_confirmation",
                    "researching",
                    "planning",
                    "executing",
                    "running",
                    "completed",
                    "failed",
                    "cancelled",
                    "blocked",
                    "assisted_executing",
                    "assisted_running",
                    "assisted_paused"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Status"
      }
    },
...
  }
}